shared type ports
