+++ apps/desktop/src/renderer/stores/ports/index.ts
@@ -1,1 +1,2 @@
-export { type DetectedPort, usePortsStore } from "./store";
+export type { DetectedPort } from "shared/types";
+export { usePortsStore } from "./store";
