+++ apps/desktop/src/shared/types/ports.ts
@@ -1,0 +1,7 @@
+export interface DetectedPort {
+	port: number;
+	paneId: string;
+	workspaceId: string;
+	detectedAt: number;
+	contextLine: string;
+}
